chore: move "bootstrap-disk-file-service" into file-explorer-agent
