+++ packages/app/src/platform/electron/file-explorer-agent/bootstrap-disk-file-service.ts
